+++ rust/stackable-cockpit/src/platform/operator/mod.rs
@@ -3,7 +3,8 @@ use std::{fmt::Display, str::FromStr};
 use semver::Version;
 use serde::Serialize;
 use snafu::{ResultExt, Snafu, ensure};
-use tracing::{info, instrument};
+use tracing::{Span, info, instrument};
+use tracing_indicatif::{indicatif_println, span_ext::IndicatifSpanExt};
 
 use crate::{
     constants::{
@@ -61,10 +62,15 @@ pub struct OperatorSpec {
 
 impl Display for OperatorSpec {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}{}", self.name, match &self.version {
-            Some(v) => format!("={v}"),
-            None => "".into(),
-        })
+        write!(
+            f,
+            "{name}{version_selector}",
+            name = self.name,
+            version_selector = match &self.version {
+                Some(v) => format!("={v}"),
+                None => "".into(),
+            }
+        )
     }
 }
 
@@ -179,6 +185,7 @@ impl OperatorSpec {
         // display for the inner type if it exists. Otherwise we gte the Debug
         // impl for the whole Option.
         version = self.version.as_ref().map(tracing::field::display),
+        indicatif.pb_show = true
     ))]
     pub fn install(
         &self,
@@ -186,6 +193,8 @@ impl OperatorSpec {
         chart_source: &ChartSourceType,
     ) -> Result<(), helm::Error> {
         info!(operator = %self, "Installing operator");
+        Span::current()
+            .pb_set_message(format!("Installing {name}-operator", name = self.name).as_str());
 
         let version = self.version.as_ref().map(|v| v.to_string());
         let helm_name = self.helm_name();
@@ -221,7 +230,7 @@ impl OperatorSpec {
     {
         match helm::uninstall_release(&self.helm_name(), namespace.as_ref(), true) {
             Ok(status) => {
-                println!("{status}");
+                indicatif_println!("{status}");
                 Ok(())
             }
             Err(err) => Err(err),