+++ rust/stackablectl/src/cmds/release.rs
@@ -20,7 +20,8 @@ use stackable_cockpit::{
     },
     xfer::{self, cache::Cache},
 };
-use tracing::{debug, info, instrument};
+use tracing::{Span, debug, info, instrument};
+use tracing_indicatif::span_ext::IndicatifSpanExt as _;
 
 use crate::{
     args::{CommonClusterArgs, CommonClusterArgsError},
@@ -185,13 +186,14 @@ impl ReleaseArgs {
     }
 }
 
-#[instrument(skip(cli, release_list))]
+#[instrument(skip(cli, release_list), fields(indicatif.pb_show = true))]
 async fn list_cmd(
     args: &ReleaseListArgs,
     cli: &Cli,
     release_list: release::ReleaseList,
 ) -> Result<String, CmdError> {
     info!("Listing releases");
+    Span::current().pb_set_message("Fetching release information");
 
     match args.output_type {
         OutputType::Plain | OutputType::Table => {
@@ -239,13 +241,14 @@ async fn list_cmd(
     }
 }
 
-#[instrument(skip(cli, release_list))]
+#[instrument(skip(cli, release_list), fields(indicatif.pb_show = true))]
 async fn describe_cmd(
     args: &ReleaseDescribeArgs,
     cli: &Cli,
     release_list: release::ReleaseList,
 ) -> Result<String, CmdError> {
     info!(release = %args.release, "Describing release");
+    Span::current().pb_set_message("Fetching release information");
 
     let release = release_list.get(&args.release);
 
@@ -300,13 +303,15 @@ async fn describe_cmd(
     }
 }
 
-#[instrument(skip(cli, release_list))]
+#[instrument(skip(cli, release_list), fields(indicatif.pb_show = true))]
 async fn install_cmd(
     args: &ReleaseInstallArgs,
     cli: &Cli,
     release_list: release::ReleaseList,
 ) -> Result<String, CmdError> {
     info!(release = %args.release, "Installing release");
+    Span::current().pb_set_message("Installing release");
+
     match release_list.get(&args.release) {
         Some(release) => {
             let mut output = cli.result();
@@ -427,12 +432,14 @@ async fn upgrade_cmd(
     }
 }
 
-#[instrument(skip(cli, release_list))]
+#[instrument(skip(cli, release_list), fields(indicatif.pb_show = true))]
 async fn uninstall_cmd(
     args: &ReleaseUninstallArgs,
     cli: &Cli,
     release_list: release::ReleaseList,
 ) -> Result<String, CmdError> {
+    Span::current().pb_set_message("Uninstalling release");
+
     match release_list.get(&args.release) {
         Some(release) => {
             release